+++ packages/documentation/src/components/datetimepicker/ExampleDatetimePicker.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { DateTimeInput } from '@indigit/vanilla-components'
+import { DateTimeInput, Input as VanillaInput } from '@indigit/vanilla-components'
 import { ref } from 'vue'
 
 const value = ref('2011-01-01 0:00:01')
@@ -12,17 +12,36 @@ const value2 = ref('2022-01-01 0:00:01')
       <!-- Regular -->
       <DateTimeInput
         v-model="value"
-      />
+        :is-dark="false"
+        :is24hr="false"
+        :minute-increment="5"
+        :is-required="true"
+        :popover="{ visibility: 'click' }"
+        mode="dateTime"
+      >
+        <template #default="{ inputValue, inputEvents, updateValue }">
+          <!--          <VanillaInput -->
+          <!--            :value="inputValue" -->
+          <!--            placeholder="Click to select your date" -->
+          <!--            v-on="inputEvents" -->
+          <!--          /> -->
+          <VanillaInput
+            :model-value="inputValue"
+            placeholder="Click to select your date"
+            v-on="inputEvents"
+          />
+        </template>
+      </DateTimeInput>
 
       <div class="flex items-center justify-center mx-auto text-center mt-2">
         <pre>{{ JSON.stringify(value) }}</pre>
       </div>
 
-      <!-- Regular -->
-      <DateTimeInput
-        v-model="value2"
-        errors="Sorry for this name is already taken"
-      />
+      <!--      &lt;!&ndash; Regular &ndash;&gt; -->
+      <!--      <DateTimeInput -->
+      <!--        v-model="value2" -->
+      <!--        errors="Sorry for this name is already taken" -->
+      <!--      /> -->
 
       <div class="flex items-center justify-center mx-auto text-center mt-2">
         <pre>{{ JSON.stringify(value2) }}</pre>
