+++ packages/vanilla-components/package.json
@@ -53,10 +53,10 @@
     "@popperjs/core": "^2.11.6",
     "@vueuse/core": "^9.2.0",
     "axios": "^1.1.2",
-    "flatpickr": "^4.6.9",
     "fuse.js": "^6.4.6",
     "libphonenumber-js": "^1.9.49",
     "lodash": "^4.17.21",
+    "v-calendar": "3.0.0-alpha.8",
     "vue": "^3.2.39"
   },
   "peerDependencies": {
